Update API routes in CoinsController for clarity
The `CoinsController` class in the `Duo.Api.Controllers` namespace has updated its HTTP route attributes. The routes for the `GetUserCoinBalance`, `SpendCoins`, and `ApplyDailyLoginBonus` methods have been changed to more descriptive names: "balance/{userId}" to "get-balance-by-user-id", "spend" to "add-spent-coins", and "dailybonus" to "add-daily-bonus". This improves the clarity of the API endpoints.